refactor(contacts): organize compact list files
