+++ features/flow/flow-contacts-list/src/components/ContactsCompactList/ContactsCompactList.web.test.tsx
@@ -7,7 +7,7 @@ import {
   mockContactWithMultipleAddresses,
 } from "@domain/entity-contact/schema.mock";
 import type { Contact } from "@domain/entity-contact";
-import { ContactsCompactList } from "./web";
+import { ContactsCompactList } from "../../web";
 
 const labels = {
   emptyAddress: "No saved addresses",
