+++ features/flow/flow-contacts-list/src/index.native.ts
@@ -1,4 +1,4 @@
 export * from "./exports";
 export * from "./ContactsListView.native";
-export * from "./ContactsCompactList.native";
+export * from "./components/ContactsCompactList/ContactsCompactList.native";
 export * from "./components/ListHeader/ContactsAddContactHeaderButton.native";
