+++ features/flow/flow-contacts-list/src/web.ts
@@ -1,3 +1,3 @@
 export * from "./index";
 export * from "./ContactsListView.web";
-export * from "./ContactsCompactList.web";
+export * from "./components/ContactsCompactList/ContactsCompactList.web";
